ti:  add send and receive methods for redundant network - add redundant info to ptp message Redundant info added to ptp message is for handling redundant network information such as HSR tag information and timestamp generated because of cut-through forwarding of a received PTP event packet.
- transport: add send and receive methods for redundant network
The send and receive methods added heandle the redundant network
information associated with the ptp messages being sent or
received. They also provide means to specify which slave port of
a redundant network interface over which a ptp message should be
sent. If cut-through forwarding is provided by underlying hardware,
these new methods also provide means to handle the transmit timestamp
associated with a received ptp message which is cut-through forwarded
by the underlying hardware.

- raw: implement redundant network send and receive methods
This patch implements the send and receive methods for the
redundant network.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Eric Ruei <e-ruei1@ti.com>